pyproject.toml
[tool.poetry]
name = "neovim-config"
version = "0.1.0"
description = ""
authors = ["Kristina Pianykh <kristinavrnrus@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
databricks-sdk = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
